pyproject.toml
[project]
name = "insightbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.73.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.1",
]
